import_module_from_file modif. + renamed module os
The function import_module_from_file was moved to
reframe/utility/__init__.py. Due to errors occuring
with respect to the os-module, the reframe module
was renamed to os_ext.
