+++ reframe/core/environments.py
@@ -3,7 +3,7 @@ import os
 
 import reframe.core.debug as debug
 import reframe.core.fields as fields
-import reframe.utility.os as os_ext
+import reframe.utility.os_ext as os_ext
 from reframe.core.exceptions import (EnvironError, SpawnedProcessError,
                                      CompilationError)
 from reframe.core.modules import get_modules_system
