+++ reframe/core/schedulers/slurm.py
@@ -4,7 +4,7 @@ import time
 from datetime import datetime
 
 import reframe.core.schedulers as sched
-import reframe.utility.os as os_ext
+import reframe.utility.os_ext as os_ext
 from reframe.core.exceptions import (SpawnedProcessError,
                                      JobBlockedError, JobError)
 from reframe.core.logging import getlogger
