+++ reframe/frontend/config.py
@@ -2,20 +2,20 @@ import os
 import collections.abc
 
 import reframe.core.debug as debug
-import reframe.utility.os as os_ext
+import reframe.utility.os_ext as os_ext
 from reframe.core.environments import Environment
 from reframe.core.exceptions import ConfigError, ReframeError, ReframeFatalError
 from reframe.core.fields import ScopedDict, ScopedDictField
 from reframe.core.launchers.registry import getlauncher
-from reframe.frontend.loader import RegressionCheckLoader
 from reframe.core.schedulers.registry import getscheduler
 from reframe.core.systems import System, SystemPartition
+from reframe.utility import import_module_from_file
 
 _settings = None
 
 def load_from_file(filename):
     global _settings
-    _settings = RegressionCheckLoader.import_module_from_file(filename)
+    _settings = import_module_from_file(filename)
     return _settings
