+++ reframe/utility/os_ext.py
