+++ unittests/test_environments.py
@@ -2,7 +2,7 @@ import os
 import unittest
 
 import reframe.core.environments as renv
-import reframe.utility.os as os_ext
+import reframe.utility.os_ext as os_ext
 import unittests.fixtures as fixtures
 from reframe.core.exceptions import CompilationError, EnvironError
 from reframe.core.modules import get_modules_system
